requirements.txt
# FastAPI
fastapi>=0.109.0,<0.111.0
uvicorn[standard]>=0.27.0,<0.31.0
python-multipart>=0.0.9,<0.1.0

# Database
sqlalchemy>=2.0.25,<2.1.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.1,<2.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<4.1.0
python-dateutil>=2.8.2,<3.0.0

# Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# Email
aiosmtplib>=3.0.0,<4.0.0
jinja2>=3.1.0,<4.0.0

# Storage (Cloudflare R2 / S3)
boto3>=1.34.0,<2.0.0
aioboto3>=12.0.0,<14.0.0
